requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy
langchain
langchain-core
langchain-community
langchain-google-genai